refactor: add pagination component to storybook
