+++ src/components/community/ArticleList/index.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { HeartBlankIcon, HeartFilledIcon } from "@components/common/icon/Heart";
 import { BookmarkBlankIcon } from "@components/common/icon/Bookmark";
 import CommonInfo from "@components/community/CommonInfo";
-import Pagination from "@components/common/Pagination";
+import Pagination from "@/stories/common/Pagination";
 import Skeleton from "@components/common/Skeleton/Skeleton";
 import useArticleList from "./useArticleList";
 import { ListDiv, Article, Text } from "./styled";
