+++ src/stories/assets/ChevronRight.tsx
